requirements-qt.txt
numpy
PyOpenGL
matplotlib
pyqt5
psutil